refactor: move failure class inside core folder
