+++ lib/core/usecases/usecase.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:fpdart/fpdart.dart';
 
-import '../../errors/failure.dart';
+import '../errors/failure.dart';
 
 abstract interface class UseCase<SuccessType, Params> {
   Future<Either<Failure, SuccessType>> call(Params params);
