+++ lib/features/auth/domain/usecases/auth_check_usecase.dart
@@ -1,7 +1,7 @@
 import 'package:fpdart/fpdart.dart';
 
 import '../../../../core/usecases/usecase.dart';
-import '../../../../errors/failure.dart';
+import '../../../../core/errors/failure.dart';
 import '../../data/models/user_model.dart';
 import '../repositories/auth_repository.dart';
