+++ lib/features/profile/domain/usecases/display_user_traning_info_usecase.dart
@@ -1,7 +1,7 @@
 import 'package:fpdart/fpdart.dart';
 
 import '../../../../core/usecases/usecase.dart';
-import '../../../../errors/failure.dart';
+import '../../../../core/errors/failure.dart';
 import '../../data/models/member_model.dart';
 import '../repositories/profile_repository.dart';
