+++ lib/features/profile/domain/usecases/update_info_member_usecase.dart
@@ -1,7 +1,7 @@
 import 'package:fpdart/fpdart.dart';
 
 import '../../../../core/usecases/usecase.dart';
-import '../../../../errors/failure.dart';
+import '../../../../core/errors/failure.dart';
 import '../../data/models/member_model.dart';
 import '../repositories/profile_repository.dart';
