+++ lib/features/publication/domain/usecases/add_publication_usecase.dart
@@ -2,7 +2,7 @@ import 'package:file_picker/file_picker.dart';
 import 'package:fpdart/fpdart.dart';
 
 import '../../../../core/usecases/usecase.dart';
-import '../../../../errors/failure.dart';
+import '../../../../core/errors/failure.dart';
 import '../entities/publication.dart';
 import '../entities/publication_type.dart';
 import '../repositories/publication_repository.dart';
